refactor(des): split SimuladorTab (663→80) em hook + componentes
Extrai os dados (3 queries + memos de faltamProximaFaixa/defaultPrazo/
compareDefault) para useSimuladorData e quebra o JSX em SimulationPanel
(estado + RPC simular), SimulationControls, SimulationResult, StarsRow e
format (fmtBRL/fmtPct). Named export SimuladorTab preservado (consumidor:
AdminDesTrimestreAtual). Comportamento 1:1; +18 testes. (warning exhaustive-deps
de `prazos` é baseline, idêntico ao original.)
