+++ src/components/des/simulador/format.ts
@@ -1,0 +1,10 @@
+// Helpers de formatação do simulador DES (puros).
+// Extraídos verbatim de src/components/des/SimuladorTab.tsx (god-component split).
+
+export const fmtBRL = (v: number | null | undefined) =>
+  v == null
+    ? "—"
+    : new Intl.NumberFormat("pt-BR", { style: "currency", currency: "BRL" }).format(Number(v));
+
+export const fmtPct = (v: number | null | undefined) =>
+  v == null ? "—" : `${Number(v).toFixed(2).replace(".", ",")}%`;
